add support for the Ruby strftime %L formatter (millisecond precision)
This adds support for the extension of the `strftime` as added within Ruby. This adds the `%L` conversion string which will give millisecond precision with a `.` character in the front. So `%H:%M:%S%L` would become `02:12:56.000`.

In addition, `go fmt`/`goimports` made some adjustments to the alignment of the code in certain places.